Setup serverless and folder struction
