+++ frontend/src/App.jsx
@@ -1,0 +1,7 @@
+import './App.css';
+
+function App() {
+  return <></>;
+}
+
+export default App;
