+++ frontend/syncHub/serverless.yml
@@ -1,0 +1,16 @@
+# "org" ensures this Service is used with the correct Serverless Framework Access Key.
+org: egestudent
+# "service" is the name of this project. This will also be added to your AWS resource names.
+service: syncHub
+
+provider:
+  name: aws
+  runtime: nodejs20.x
+
+functions:
+  hello:
+    handler: handler.hello
+    events:
+      - httpApi:
+          path: /
+          method: get
